requirements.txt
stellar-sdk>=11.0.0
aiohttp>=3.8.0
pyyaml>=6.0
python-dotenv>=1.0.0
streamlit>=1.28.0
pytest>=7.0.0
plotly>=5.0.0
pandas>=1.4.0
scipy>=1.9.0
numpy>=1.23.0
requests>=2.28.0
openai>=1.0.0
